move thirdparty folder
